Integrate with CI services
We want to ensure that any changes we make won't break the project, follow current best practise, and are secure. This is why we always integrate our projects with various online CI services right from the very start.

Though not all CI services require changes to the code, with this change we are signifying integrations with

- [Travis CI](https://travis-ci.com/DEFRA/defra-ruby-address)
- [CodeClimate](https://codeclimate.com/github/DEFRA/defra-ruby-address)
- [Hakiri](https://hakiri.io/github/DEFRA/defra-ruby-address/master)
- [Dependabot](https://app.dependabot.com/accounts/DEFRA/repos/211942150)