Add Response object
Irrespective of what address lookup is used, the intention is that all callers will receive the same instance type of `Response` with its set API of `successful?`, `results` and `error`.

`Response.results` will also always be an array of JSON hashes, however the format of each will be dependent on which address lookup is used (once we implement it).

It will then be up to the calling app to use the data as it needs.