Add OS Places Address Lookup service
The [os-places-address-lookup](https://github.com/DEFRA/os-places-address-lookup) is the OS Places facade used by the Waste Carriers Registration (WCR) service.

This is one of the facades we'd like this gem to provide a uniform API to, and we've picked it as the first to add to the gem.

Within this change we add a base service class which will handle all logic for making the request going forward. The individual facade service classes really only need to provide the correctly formatted URL.