Handle no match errors
The service we are replacing interpreted `RestClient::BadRequest` errors as there being no match. Our tests show though that we'll get a `RestClient::BadRequest` for a number of reasons. So just assuming `BadRequest` == `no_match` doesn't really cut it.

However the lookup doesn't give us much to go on other than the error message itself. It seems if we pass an unrecognised postcode we get the message `"Parameters are not valid"`.

So this change adds some additional code to try and cater for this in the service, so we replicate what the existing one does for waste carriers.