First step in repurposing the base service
Having amended the base service to handle how the OS Places Address Lookup responds to no matches, it became clear trying to have a single base address service the others inherit from is not going to be possible (or at least pretty messy).

So this is the first step in altering what it does. We rename the file here to just be `base_service`, and in the next change we'll move the functionality into the OS places service.

`base_service.rb` itself will follow the pattern we use in our actual apps, we're it'll just be used to define that all services have a `run()` method.