Remove non-param run() from base service
This was causing test coverage to drop because we had no service exercising a parameterless `run()`. On consideration we couldn't think of when we would try to make an address lookup without a postcode, so rather than add a meaningless unit test to up the code coverage, we refactored the `BaseService` class.