Fix no match error for other errors
Coverage was still a little low and on further inspection it was because when checking if a failed response was a no match, we were actually generating a new error which was short circuiting the method.

So this fixes the issue in `no_match_error?` and ups our test coverage to 100% again.