Add OS Places Address Lookup service (#5)
The [os-places-address-lookup](https://github.com/DEFRA/os-places-address-lookup) is the OS Places facade used by the Waste Carriers Registration (WCR) service.

This is one of the facades we'd like this gem to provide a uniform API to, and we've picked it as the first to add to the gem.

Within this change we add a base service class to ensure all the lookup services expose a `run()` method, and the new `OsPlacesAddressLookupService` class.

Ideally we would have shifted as much of the request logic into the base, but when working on this change it became clear that the way this address lookup works and responds is likely to be very different to the Address Base Facade so we avoided doing that at this stage.

Also worth noting we do not expect the API's of the services we are calling to change. If they do we'll add a new service to this gem. So using VCR and having to re-record cassettes seems an unnecessary overhead. So in this change we've removed VCR, and going forward all tests will use WebMock to stub the requests instead.