Add additional config attributes
The EA Address Facade expects a `client-id` and `key` to be passed as query params when making a request. In practise the default config is to set these to `0` and `'client1'`, and we're not aware of a service that has changed them 😀.

So as they are configurable we add them to our config, but as we never change them we default their values.