Add ea address facade service
Working with a slightly better API we don't have to do anything clever like trying to determine the actual response from the message buried in the error message.

Also due to the size of the response, felt it was better to start saving the body in json files rather than bloat the spec with the detail. Hence the introduction of fixtures.