Refactor EA Address Facade services to reduce dup.
There was obvious duplication between the EA Address Facade services which CodeClimate has flagged.

This is an attempt at a solution which avoids inheritence in favour of composition.

It adds a new object which handles forming the url, making the request to the address lookup, and returning a result. Each service then just instantiates an instance of the object when needed.