Add Addressbase facade version 1.1 service (#7)
The [EA Address Facade Service](https://github.com/DEFRA/ea-address-facade) comes in 2 flavours; version 1 and 1.1.

Currently [Waste Exemptions](https://github.com/DEFRA/ruby-services-team/tree/master/services/wex) and [Flood Risk Activity Exemptions](https://github.com/DEFRA/ruby-services-team/tree/master/services/frae) are on version 1. However, we would like to move them to version 1.1 in the near future, as it has enhanced security and rate-limiting.

So to prepare, this change adds a new service to the gem that works with version 1.1 of the facade's API.

## Refactor EA Address Facade services to reduce dup.

There was obvious duplication between the EA Address Facade services once we'd implemented v1.1.

So also within this change, we created a new `EaAddressFacadeRequest` class to contain the logic used by both. We went for this solution to avoid inheritance in favour of composition.

Each service then just instantiates an instance of the object when needed.