Update changelog config in Rakefile
The current confg was added by simply copying what we have done in countless other projects.

When we did those, as long as the env var `CHANGELOG_GITHUB_TOKEN` was set we simply needed to call `bundle exec rake changelog` in the root of the project and the GitHub username and project seemed to be automatically determined.

However when coming to create the initial changelog for this project, this no longer appears to be the case. It seems something has changed in v1.15.0 of [github-changelog-generator](https://github.com/github-changelog-generator/github-changelog-generator) that means these values need to be provided.

So rather than having to remember them each time when running the command, this change uses the ability to define them in the changelog's config instead.