Fix rubocop cop rename issue (#8)
We spotted that automated builds on master by Travis-CI had started failing.

The issue is that rubocop has released a new version (0.77) and it looks like it has changed the name of the cop which checks suppressed exceptions.

We suppress them because they are environment specific. We don't want the `Rakefile` to fail if the environment is set to production, just because we reference development only libraries in it.